// src/core/types/cards.types.ts
export interface ICompanyCardProps {
  header: string;
  text: string;
  background: string;
  textColor: string;
  headerColor: string;
  image: any;
}

export interface IVehicleCardProps {
  name: string;
  price: number;
  image: any;
}

export interface IInstructionCardProps {
  header: string;
  text: string;
}